Module Tezos_p2p

module P2p : sig ... end

Tezos P2p layer - Dynamic overlay network of authenticated peers.

module P2p_acl : sig ... end

This module implements four Access Control Lists:

module P2p_answerer : sig ... end

An Answerer.t is a set of callback functions, parameterized by conn_info record. The conn_info records contains values useful for the callback functions to perform their task, and known after the connection is set up.

module P2p_conn : sig ... end

Type of a connection to a peer, parametrized by the type of messages exchanged as well as meta-information associated to a peer and a connection. It wraps a P2p_socket.t, adding meta-information and data-structures describing a more fine-grained logical state of the connection. It also set up an answering worker that responds to the messages P2p_message.t using the callback functions of a P2p_answerer.t.

module P2p_connect_handler : sig ... end

This module manages incoming accept and outgoing connections connect.

module P2p_directory : sig ... end

module P2p_discovery : sig ... end

Local peer discovery.

module P2p_events : sig ... end

module P2p_fd : sig ... end

This module defines a type t which wraps a file descriptor. Most functions simply call the underlying file descriptor function and generate logs with prefix "p2p.fd".

module P2p_io_scheduler : sig ... end

Scheduling of I/O operations over file descriptors.

module P2p_maintenance : sig ... end

P2P maintenance worker.

module P2p_message : sig ... end

This module defines the messages of the P2p layers.

module P2p_params : sig ... end

Types of the abstract data types (ADT) which parameterize the P2p layer.

module P2p_peer_state : sig ... end

module P2p_point_state : sig ... end

module P2p_pool : sig ... end

This module maintains several pools of points/peers needed by the P2P layer.

module P2p_protocol : sig ... end

This module defines constructor for the private and default P2p_answerer.t. Both pass messages `Message msg` to the upper-layer. The private answerer ignore all other messages.

module P2p_socket : sig ... end

Typed and encrypted connections to peers.

module P2p_trigger : sig ... end

This module defines some condition values for inter modules synchronization.

module P2p_welcome : sig ... end

Welcome worker.