Module Tezos_p2p.P2p_conn

Type of a connection to a peer, parametrized by the type of messages exchanged as well as meta-information associated to a peer and a connection. It wraps a P2p_socket.t, adding meta-information and data-structures describing a more fine-grained logical state of the connection. It also set up an answering worker that responds to the messages P2p_message.t using the callback functions of a P2p_answerer.t.

type ('msg, 'peer, 'conn) t

val create : conn:('msg P2p_message.t, 'conn) P2p_socket.t -> point_info:('msg, 'peer, 'conn) t P2p_point_state.Info.t option -> peer_info:(('msg, 'peer, 'conn) t, 'peer, 'conn) P2p_peer_state.Info.t -> messages:(int * 'msg) Tezos_stdlib.Lwt_pipe.t -> canceler:Lwt_canceler.t -> greylister:(unit -> unit) -> callback:'msg P2p_answerer.t -> Tezos_base.Network_version.t -> ('msg, 'peer, 'conn) t

val peer_id : ('msg, 'peer, 'conn) t -> Tezos_base.P2p_peer.Id.t

val private_node : ('msg, 'peer, 'conn) t -> bool

private_node t returns 'true' if the node associated to this connection is in private mode

val trusted_node : ('msg, 'peer, 'conn) t -> bool

trusted_node t returns 'true' if the node associated to this connection is trusted

val info : ('msg, 'peer, 'conn) t -> 'conn Tezos_base.P2p_connection.Info.t

val local_metadata : ('msg, 'peer, 'conn) t -> 'conn

val remote_metadata : ('msg, 'peer, 'conn) t -> 'conn

val stat : ('msg, 'peer, 'conn) t -> Tezos_base.P2p_stat.t

stat t is a snapshot of current bandwidth usage for t.

val read : ('msg, 'peer, 'conn) t -> ('msg, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

read t returns a message popped from t's app message queue, or fails with Connection_closed.

val is_readable : ('msg, 'peer, 'conn) t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

is_readable t returns when there is at least one message ready to be read.

val write_swap_request : ('msg, 'peer, 'conn) t -> Tezos_base.P2p_point.Id.t -> Tezos_base.P2p_peer.Id.t -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result

val write_swap_ack : ('msg, 'peer, 'conn) t -> Tezos_base.P2p_point.Id.t -> Tezos_base.P2p_peer.Id.t -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result

val write_bootstrap : ('msg, 'peer, 'conn) t -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result

val write : ('msg, 'peer, 'conn) t -> 'msg -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

write t msg is P2p_socket.write t' msg where t' is the internal P2p_socket.t inside t.

val write_sync : ('msg, 'peer, 'conn) t -> 'msg -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

write_sync t msg is P2p_socket.write_sync t' msg where t' is the internal P2p_socket.t inside t.

val write_now : ('msg, 'peer, 'conn) t -> 'msg -> (bool, Tezos_error_monad.TzCore.error list) Stdlib.result

write_now t msg is P2p_socket.write_now t' msg where t' is the internal P2p_socket.t inside t.

val equal_sock : ('msg, 'peer, 'conn) t -> ('msg, 'peer, 'conn) t -> bool

val disconnect : ?wait:bool -> ('msg, 'peer, 'conn) t -> unit Lwt.t

val close : ('msg, 'peer, 'conn) t -> unit Lwt.t