Module Tezos_p2p.P2p_fd

This module defines a type t which wraps a file descriptor. Most functions simply call the underlying file descriptor function and generate logs with prefix "p2p.fd".

type t

val id : t -> int

id t returns a unique, positive, identifier for t. Identifiers are generated sequentially at creation time.

val read : t -> Stdlib.Bytes.t -> int -> int -> int Lwt.t

val close : t -> (unit, Tezos_error_monad.TzCore.error list) Stdlib.result Lwt.t

val write : t -> Stdlib.Bytes.t -> unit Lwt.t

val socket : Lwt_unix.socket_domain -> Lwt_unix.socket_type -> int -> t Lwt.t

val connect : t -> Lwt_unix.sockaddr -> unit Lwt.t

val accept : Lwt_unix.file_descr -> (t * Lwt_unix.sockaddr) Lwt.t

module Table : sig ... end