Module P2p_io_scheduler.Internal

Module for testing only, do not use in production.

val destruct_buffer : buffer -> int * int * Stdlib.Bytes.t

destruct_buffer buf returns the pos, len, and buf values of the given buffer. See mk_buffer.