Module Tezos_p2p__P2p_acl

This module implements four Access Control Lists:

IP greylist is a set of banned IP addresses automatically added by the P2P layer.

peer_id greylist is a set of banned peers ids automatically added by the P2P layer.

IP blacklist is a set of IP addresses manually added by the node admin.

peers blacklist is a set of peers ids manually added by the node admin.

IP greylists use a time based GC to periodically remove entries from the table, while peer_id greylists are built using an LRU cache, where the least-recently grey-listed peer is evicted from the table when adding a new banned peer to a full cache. Other tables are user defined and static.