Module Tezos_p2p__P2p_acl.PeerGreylist

val add : t -> Tezos_base__TzPervasives.P2p_peer.Id.t -> unit

add t peer adds peer to the peer greylist. It might potentially evict the least-recently greylisted peer, if the grey list is full. If peer was already in the list, it will become the most-recently greylisted, thus ensuring it is not evicted unfairly soon.

val mem : t -> Tezos_base__TzPervasives.P2p_peer.Id.t -> bool

mem t peer returns true iff peer is greylisted.

val list : t -> Tezos_base__TzPervasives.P2p_peer.Id.t list

list t: return the list peers added to the greylist