Module Tezos_p2p__P2p_answerer

An Answerer.t is a set of callback functions, parameterized by conn_info record. The conn_info records contains values useful for the callback functions to perform their task, and known after the connection is set up.

The callback functions are called when the node receives `P2p_messages.t` messages. The parameters are the values carried by the message, and a request_info record that contains values pertaining to the connection that may change during the life of the connection.