Module Tezos_p2p__P2p_discovery

Local peer discovery.

This module manages the discovery of local peers through UDP broadcasting. It is composed of two workers:

The sender worker whose role is to broadcast discovery messages.

The answer worker whose role is to listen discovery messages and register new peers in the current pool. Discovery messages are composed of an arbitrary key, the listening port and the peer id of the current peer.