Module Tezos_p2p__P2p_params

Types of the abstract data types (ADT) which parameterize the P2p layer.

Many types used in the P2p layer are parameterized by three type parameters:

'msg: type of messages exchanged between peers

'peer_meta: type of the metadata associated with peers (score, etc.)

'conn_meta: type of the metadata associated with connections

These types are kept abstract from the P2p layer. It can only operate on them via a set of functions packed in a "configuration" record passed by the upper layer (see also P2p and P2p.create).

This module defines the type of these configuration records.