Module Tezos_p2p__P2p_welcome

Welcome worker.

Accept incoming connections and add them to the pool.

type t

Type of a welcome worker.

val create : ?⁠addr:Tezos_base__TzPervasives.P2p_addr.t -> backlog:int -> ('msg, 'meta, 'meta_conn) Tezos_p2p.P2p_connect_handler.t -> Tezos_base__TzPervasives.P2p_addr.port -> t Lwt.t

create ?addr ~backlog pool port returns a running welcome worker adding connections into pool listening on addr:port. backlog is passed to Lwt_unix.listen.

val activate : t -> unit

activate t start the worker that will accept connections

val shutdown : t -> unit Lwt.t

shutdown t returns when t has completed shutdown.